% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{as.ic_data}
\alias{as.ic_data}
\alias{is.ic_data}
\title{Convert foreign object to an ic data object}
\usage{
as.ic_data(...)

is.ic_data(object)
}
\arguments{
\item{...}{Elements to be combined into an \code{ic.df} object.}

\item{object}{Any \code{R} object.}
}
\value{
An object of type \code{ic.df}.
}
\description{
Functions to check if an object is an ic data object, or coerce
it to one if possible.
}
\details{
Core \code{ic.df} data elements are expected. Specifically, columns
of data for 'region', 'country', 'time', 'vaccine', 'coverage', and
'source' (optionally 'dose' and 'population') are required in this order.
The names of these elements from \code{...} are passed to \code{ic_data}
and used along with the default settings.
}
\examples{
\dontrun{
# convert data.frame to an imcover data frame
ic_df <- as.ic_data(df)
}

}
\seealso{
\code{\link[imcover]{ic_data}}
}
