% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{download_wuenic}
\alias{download_coverage}
\alias{download_survey}
\alias{download_denom}
\title{Download WUENIC data files}
\usage{
download_wuenic(
  destfile,
  url,
  use_cache = TRUE,
  quiet = FALSE,
  attempts = 3,
  mode = "wb",
  return_ic = FALSE,
  add_region = "who",
  ...
)

download_coverage(
  destfile,
  url,
  use_cache = TRUE,
  quiet = FALSE,
  attempts = 3,
  mode = "wb",
  return_ic = TRUE,
  add_region = "who",
  ...
)

download_survey(
  destfile,
  url,
  use_cache = TRUE,
  quiet = FALSE,
  attempts = 3,
  mode = "wb",
  return_ic = TRUE,
  reduce = TRUE,
  minSample = 300,
  adjust = TRUE,
  adjVacc = c("DTP", "PCV"),
  add_region = "who",
  ...
)

download_denom(
  destfile,
  url,
  use_cache = TRUE,
  quiet = FALSE,
  attempts = 3,
  mode = "wb",
  ...
)
}
\arguments{
\item{destfile}{Optional. Character string with the name of where the
downloaded file is saved.}

\item{url}{Optional. A character string naming the URL of the resource to
download. Default is the .xlsx from the WHO immunization data website.}

\item{use_cache}{Logical. Look for an already downloaded file is the
\code{destfile} location? Default is \code{TRUE}.}

\item{quiet}{Should status messages and progress bars be hidden? Default is
\code{FALSE}.}

\item{attempts}{integer of the number of attempts to download the file.
Default is 3.}

\item{mode}{a character string specifying the write mode. 'wb' is most common
to write binary files.}

\item{return_ic}{Should an object of type \code{ic.df} be returned? Default
is \code{TRUE}.}

\item{add_region}{Optional. Character string for the type of region code to
add based on the ISO3 code of the country. Options are: 'who' or 'm49'.}

\item{...}{additional arguments to be passed on to \code{download.file}.}

\item{reduce}{For surveys only, should records with insufficient coverage
be dropped? Default is \code{TRUE}.}

\item{minSample}{For surveys only, what is the minimum sample size to keep
survey records? Default is 300.}

\item{adjust}{For survey only, should a recall-bias adjustment be applied?
Default is \code{TRUE}.}

\item{adjVacc}{Character vector of vaccines to adjust when \code{adjust} is
\code{TRUE}. Default is "DTP" and "PCV".}
}
\value{
The functions return a \code{data.frame} while
\code{download_coverage} and \code{downlaod_survey} return objects of type
\code{ic.df} unless the option \code{return_ic} is \code{FALSE}.
}
\description{
Retrieve the input data files and WUENIC coverage estimates with further
processing of \code{ic.df} objects.
}
\seealso{
\code{\link[imcover]{ic_data}}
}
