% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_yovi}
\alias{filter_yovi}
\alias{filter_yovi.ic.df}
\alias{filter_yovi.icfit}
\alias{filter_yovi.iclist}
\title{Filter ic data records based on year of vaccine introduction
Exclude ic data records for years when a vaccine had not been fully
introduced in a country.}
\usage{
filter_yovi(X, vaccine = NULL, yovi = imcover::get_yovi(), na.rm = FALSE)

\method{filter_yovi}{ic.df}(X, vaccine = NULL, yovi = imcover::get_yovi(), na.rm = FALSE)

\method{filter_yovi}{icfit}(X, vaccine = NULL, yovi = imcover::get_yovi(), na.rm = FALSE)

\method{filter_yovi}{iclist}(X, vaccine = NULL, yovi = imcover::get_yovi(), na.rm = FALSE)
}
\arguments{
\item{X}{An object of class \code{ic.df} or \code{icfit}.}

\item{vaccine}{Vector of vaccine codes as characters. Default is \code{NULL}
which includes all vaccines in object \code{X}.}

\item{yovi}{Table of the year of vaccine introduction. See details. Default
value comes from \code{imcover::get_yovi()}.}

\item{na.rm}{Should records which do not have any data for year of
introduction be removed? Default is \code{FALSE} to keep \code{NA} records
in \code{X}.}
}
\value{
An object with type of \code{X} with records filtered by the year of
vaccine introductions.
}
\description{
Filter ic data records based on year of vaccine introduction
Exclude ic data records for years when a vaccine had not been fully
introduced in a country.
}
\details{
The \code{yovi} argument can take a user-defined table of values.
This must be a \code{data.frame} object with column names 'ISO3code',
'vaccine' and 'year_introduced'.
}
\seealso{
\code{\link[imcover]{get_yovi}}
}
