% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_region}
\alias{get_region}
\title{Get regional groupings for country codes
Retrieve a list of the region based on the ISO3c alpha codes.}
\usage{
get_region(country = NULL, type = "who")
}
\arguments{
\item{country}{Vector of country ISO3 codes as characters. Default is \code{NULL}
which returns all regions in the dataset.}

\item{type}{The type of region grouping to return. The default \code{'who'}
returns World Health Organisation codes. The other option is 'M49' for UN
Stats Division region groupings.}
}
\value{
A character vector of region codes of length equal
\code{length(country)} if \code{country} is specified or a \code{data.frame} of
the region table if \code{country} is \code{NULL}.
}
\description{
Get regional groupings for country codes
Retrieve a list of the region based on the ISO3c alpha codes.
}
\details{
Data source of M49 regions:
\link{https://unstats.un.org/unsd/methodology/m49/overview/}
}
