% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_yovi}
\alias{get_yovi}
\title{Get year of introduction for vaccines
Retrieve a list of the year when a vaccine was introduced to the entire
country.}
\usage{
get_yovi(vaccine = NULL)
}
\arguments{
\item{vaccine}{Vector of vaccine codes as characters. Default is \code{NULL}
which returns all vaccines in the dataset.}
}
\description{
Get year of introduction for vaccines
Retrieve a list of the year when a vaccine was introduced to the entire
country.
}
\details{
Data source of vaccine schedules:
\link{https://www.who.int/immunization/monitoring_surveillance/data/year_vaccine_introduction.xlsx}
}
