% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic.df}
\alias{ic.df}
\alias{[.ic.df}
\title{Access ic data}
\usage{
\method{[}{ic.df}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{Object of class \code{ic.df} from which to extract element(s) or in
which to replace element(s).}

\item{i, j, ...}{Indices specifying the elements to extract or replace. Numeric
or character vector or empty (missing) or NULL.}

\item{drop}{Can core ic attribute columns be modified? Default is
\code{FALSE}.}
}
\description{
Operators to extract or replace parts of ic data.
}
\details{
\code{[.ic.df} will generally follow the behaviour of
\code{data.frame}; however, when \code{drop = FALSE} the core 'ic'
attributes will always be preserved and will always return an object of
class \code{ic.df}.
}
\seealso{
\code{ic_data}
}
