% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{ic_adjust}
\alias{ic_adjust}
\alias{ic_ratio}
\title{Adjust 'ic' coverage data}
\usage{
ic_adjust(
  X,
  ratio_adj = FALSE,
  numerator = "DTP3",
  denominator = "DTP1",
  coverage_adj = TRUE
)

ic_ratio(X, numerator = "DTP3", denominator = "DTP1")
}
\arguments{
\item{X}{Object of class \code{ic.df} to adjust}

\item{ratio_adj}{Logical. Should multi-dose vaccine coverages be adjusted by
their ratio when coverage > 100\%? Default is \code{FALSE}.}

\item{numerator}{Character vector of vaccine identifiers (with dose numbers)
to calculate the ratio adjustment. Default is 'DTP3'.}

\item{denominator}{Character vector of vaccine identifiers (with dose
numbers) to calculate the ratio adjustment. Default is 'DTP1'.}

\item{coverage_adj}{Logical. Should coverage percentage be changed to be >0\%
and <100\%? Default is \code{TRUE}.}
}
\description{
Make adjustments to vaccination coverage to ensure reported coverages
preserve the ratio of multi-dose vaccines and that coverage is >0\% and <100\%.
}
\details{
This function provides minimal adjustments to immunisation data to
ensure that the coverages are suitable for statistical modelling. Ratio
adjustments help to ensure that multi-dose vaccines report a coverage that
is less-than-or-equal-to the coverage of the first dose.

Coverage adjustment accounts for inaccurate target populations that result
is coverages of >100\%. By maintaining a coverage between 0 and 100\% a
logistic transformation can be used in the statistical model.
}
\seealso{
\code{\link[imcover]{ic_data}}
}
