% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic_core}
\alias{ic_core}
\title{List core attributes for ic data
Retrieve the names of the core attributes that should be defined in
\code{ic.df} data.}
\usage{
ic_core()
}
\description{
List core attributes for ic data
Retrieve the names of the core attributes that should be defined in
\code{ic.df} data.
}
