% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic_data}
\alias{ic_data}
\alias{ic_data.ic.df}
\alias{ic_data.data.frame}
\title{Create an ic data object}
\usage{
ic_data(
  X,
  region = "region",
  country = "code",
  time = "year",
  vaccine = "antigen",
  coverage = "coverage",
  source = "coverage_category",
  dose = "doses",
  population = "target_number",
  dropCols = FALSE,
  ...
)

\method{ic_data}{ic.df}(
  X,
  region = "region",
  country = "code",
  time = "year",
  vaccine = "antigen",
  coverage = "coverage",
  source = "coverage_category",
  dose = "doses",
  population = "target_number",
  dropCols = FALSE,
  ...
)

\method{ic_data}{data.frame}(
  X,
  region = "region",
  country = "code",
  time = "year",
  vaccine = "antigen",
  coverage = "coverage",
  source = "coverage_category",
  dose = "doses",
  population = "target_number",
  dropCols = FALSE,
  ...
)
}
\arguments{
\item{X}{An \code{R} object to process and convert to ic data.}

\item{region, country, time, vaccine, coverage, dose, population}{Character of the
name within \code{X} which defines the core value for ic data.}

\item{dropCols}{Should other columns in \code{X} be dropped if they are not
core attributes? Default is \code{FALSE} to retain all data from \code{X}.}

\item{...}{Additional arguments. Not currently used.}
}
\value{
An object of class \code{ic.df} which extends \code{data.frame}-like
objects with attributes to location and preserve core data elements for
immunisation coverage.
}
\description{
Create an ic data object, which extends \code{data.frame}-like objects by
defining consistent attributes for required data elements.
}
\details{
\code{ic_data} is the core function to create a properly formed and
processed dataset for immunisation coverage modelling. In particular it
requires some core data elements are present:
\itemize{
\item{'region'}{ Larger grouping. Used to apply separate models.}
\item{'country'}{ Aggregate grouping for data records, e.g. country. Country
can be defined by more than column for a nested hierarchy of units, largest
to finest grouping. Default name is 'code' referring to an ISO3 code.}
\item{'time'}{ Defines the time period of immunisation records, typically
an integer year. Default name is 'year'.}
\item{'vaccine'}{ Code to identify the vaccine records, e.g. 'DTP1'.
Default name is 'antigen'.}
\item{'coverage'}{ Pre-calculated coverage percentage. Default name is
'coverage'.}
\item{'source'}{ Character identifying the source of the data. Default is
'coverage_category'.}
}
If 'coverage' is not included, then two other elements are required to be
specified in order for percent coverage to be calculated. Else, these are
optional elements for ic data.
\itemize{
\item{'dose'}{ Number of vaccine doses administered. Default name is
'doses'.}
\item{'population'}{ Total target population for the vaccine. Default name
is 'target_number.}
}
}
\seealso{
\code{\link[imcover]{ic_expand}}, \code{ic_validate}
}
