% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic_expand}
\alias{ic_expand}
\title{Expand ic data to include all potential year.}
\usage{
ic_expand(X, min = 1999, max, na.remove = TRUE)
}
\arguments{
\item{X}{Object of class \code{ic.df} to expand.}

\item{min}{Start of the potential time period. Default is 1999.}

\item{max}{End of the potential time period.}

\item{na.remove}{Should group-vaccine combinations which have no coverage
data be removed? Default is \code{TRUE}}
}
\description{
Expand the time dimension of an ic dataset to include the full set of
potential observations.
}
\details{
If \code{max} is missing, the maximum time value observed in the
dataset will be used.
}
