% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{ic_fit_single}
\alias{ic_fit_single}
\title{Multi-source immunisation coverage model with Stan}
\usage{
ic_fit_single(X, region = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{X}{Object of \code{ic.df} for analysis}

\item{region}{Logical. Should region-specific models be generated? Default
is \code{TRUE}.}

\item{verbose}{Logical. Should messages be displayed? Default is \code{TRUE}.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{icfit} for single region models, or
\code{iclist} for multiple regions. The \code{icfit} object contains a
\code{stanfit} object returned by \code{rstan::sampling} with the fitted
object along with the posterior samples, data used in the model, and other
attributes related to the fitting. An \code{iclist} object is an extension
of a list containing multiple \code{icfit} objects.
}
\description{
Sampling from a single likelihood model using a random effect
for data source.
}
\details{
The current version of \code{imcover} does not allow for users to
modify the prior distributions in the single likelihood model. Future
versions of the package may enable this functionality. Lambda is assigned a
prior distribution of \code{normal(0, 1)} and the sigma parameters are
assigned half-Cauchy priors with a scale of 2.
}
