% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ic_plot}
\alias{ic_plot}
\alias{ic_plot.icfit}
\alias{ic_plot.iclist}
\title{Plot ic fit objects}
\usage{
ic_plot(X, observed, prob = c(0.025, 0.975), vaccine, filter_yovi = TRUE)

\method{ic_plot}{icfit}(X, observed, prob = c(0.025, 0.975), vaccine, filter_yovi = TRUE)

\method{ic_plot}{iclist}(X, observed, prob = c(0.025, 0.975), vaccine, filter_yovi = TRUE, yovi)
}
\arguments{
\item{X}{Object of type \code{icfit} or \code{iclist}}

\item{observed}{Object of type \code{ic.df}}

\item{prob}{Numeric vector of length 2 for probabilities with values in \link{0,1}
to be used to calculate the upper and lower credible intervals.}

\item{vaccine}{Optional. Character vector to subset which vaccines to plot}

\item{filter_yovi}{Logical. Should the estimates be filtered by year of
vaccine introduction?}
}
\value{
A plot of the coverage estimates extracted from the fit, including
credible intervals.
}
\description{
Provides methods to plot estimated time trends of vaccine coverage for
countries.
}
\details{
If observed data are provided, they will be overlaid on the
estimates.
}
