% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalise.R
\name{ic_regional}
\alias{ic_regional}
\alias{ic_regional.icfit}
\alias{ic_regional.iclist}
\title{Regional immunisation coverage}
\usage{
ic_regional(
  X,
  denom,
  stat = c("mean", "median", "sd", "quantile"),
  prob = c(0.025, 0.25, 0.5, 0.75, 0.975),
  filter_yovi = TRUE
)

\method{ic_regional}{icfit}(
  X,
  denom,
  stat = c("mean", "median", "sd", "quantile"),
  prob = c(0.025, 0.25, 0.5, 0.75, 0.975),
  filter_yovi = TRUE
)

\method{ic_regional}{iclist}(
  X,
  denom,
  stat = c("mean", "median", "sd", "quantile"),
  prob = c(0.025, 0.25, 0.5, 0.75, 0.975),
  filter_yovi = TRUE
)
}
\arguments{
\item{X}{Object of \code{icfit} or \code{iclist} from model fitting}

\item{denom}{Population denominator data. Default is downloaded from WHO.}

\item{stat}{Character vector of summary statistics to apply.}

\item{prob}{Numeric vector of probabilities with values in \link{0,1} to be used
for \code{stat} "quantile".}

\item{filter_yovi}{Logical. Should the estimates be filtered by year of
vaccine introduction?}
}
\value{
\code{data.frame} object with regional summaries
}
\description{
Calculates population-weighted regional estimates from \code{icfit} objects.
}
