% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{ic_survey}
\alias{ic_survey}
\alias{ic_survey.ic.df}
\alias{ic_survey.data.frame}
\title{Create an ic data object}
\usage{
ic_survey(
  X,
  ...,
  survey = "surveyNameEnglish",
  sample = "Sample_Size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300
)

\method{ic_survey}{ic.df}(
  X,
  ...,
  dropCols = FALSE,
  survey = "surveyNameEnglish",
  sample = "Sample_Size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300
)

\method{ic_survey}{ic.df}(
  X,
  ...,
  dropCols = FALSE,
  survey = "surveyNameEnglish",
  sample = "Sample_Size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300
)

\method{ic_survey}{data.frame}(
  X,
  ...,
  dropCols = FALSE,
  survey = "surveyNameEnglish",
  sample = "Sample_Size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300
)
}
\arguments{
\item{X}{Any object with immunisation coverage survey data to convert into an
ic object.}

\item{...}{Additional parameters for names of core data elements to pass on
to \code{ic_data}. See details and \code{ic_data}.}

\item{survey, sample, evidence, validity}{Character of the name within \code{X}
which defines the core values for ic survey data.}

\item{reduce}{Should country-vaccine combinations which have insufficient
coverage data be removed? Default is \code{TRUE}.}

\item{minSample}{Numeric. Minimum reported sample size needed to keep
coverage records. Default is 300.}
}
\value{
An object of class \code{ic.df} which extends \code{data.frame}-like
objects with attributes to location and preserve core data elements for
immunisation coverage.
}
\description{
Create an ic data object from survey data on immunisation
coverage.
}
\details{
\code{ic_survey} is is an extension to \code{ic_data} and is used to
create a properly formed and processed dataset for immunisation coverage
modelling based on survey datasets.

The parameters \code{min} and \code{max} are only needed when \code{expand =
TRUE}. If \code{max} is missing, the maximum time value observed in the
dataset will be used.
}
\seealso{
\code{\link[imcover]{ic_data}}
}
