% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{ic_to_stan}
\alias{ic_to_stan}
\title{Prepare 'ic' data for Stan model}
\usage{
ic_to_stan(X, make_index = TRUE)
}
\arguments{
\item{X}{Object of \code{ic.df} to be converted}

\item{make_index}{Logical. Should indices for e.g. 'country' be created?}
}
\value{
A list with named objects suitable for use as Stan data.
Default is \code{TRUE}.
}
\description{
Prepare 'ic' data for Stan model
}
\keyword{internal}
