% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ic_update}
\alias{ic_update}
\title{Update ic data to adjust denominators.}
\usage{
ic_update(admin, denom, official, survey, dtpCheck = TRUE, limit = 10)
}
\arguments{
\item{admin}{Object of class \code{ic.df} to containing administrative
estimates of coverage.}

\item{denom}{Object of class \code{ic.df} containing official WUENIC
denominators.}

\item{official}{Object of class \code{ic.df} containing official country
estimates of coverage.}

\item{survey}{Object of class \code{ic.svy} containing annual survey
estimates of coverage.}

\item{dtpCheck}{Should dose 3 and dose 1 of DTP vaccines be compared for each
country and time? If dose 3 coverage exceeds dose 1, these records will
marked \code{NA}. Default is \code{TRUE}.}

\item{limit}{Maximum acceptable difference between survey and administrative
coverage records. When the difference exceeds the limit, the survey
estimate is preferred. Default is 10 percent.}
}
\description{
Provide data checking and validation by comparing across official estimates,
administrative records, and survey estimates of immunisation coverage.
}
\seealso{
\code{\link[imcover]{ic_data}}
}
