% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcover.R
\docType{package}
\name{imcover-package}
\alias{imcover-package}
\title{imcover: An R package for spatio-temporal immunisation coverage modelling}
\description{
\link{Add some description here.}
}
\section{Credits}{

This work was undertaken by members of the WorldPop Research Group at the
University of Southampton(\url{https://www.worldpop.org/}).
}

