% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{log_lik.icfit}
\alias{log_lik.icfit}
\alias{log_lik}
\title{Extract  posterior sample of the pointwise log-likelihood from a icfit object.}
\usage{
\method{log_lik}{icfit}(X, permuted = TRUE, ...)
}
\arguments{
\item{X}{object of type \code{icfit}}

\item{permuted}{Logical. Should the draws from each chain be permuted and
merged?. Default is \code{TRUE} to merge.}

\item{...}{Not currently used.}
}
\value{
An S x N matrix containing pointwise log-likelihood samples, where S
is the number of samples and N is the number of observations in the data.
If \code{permuted} is \code{FALSE} then an S x N x R array is returned,
where R is the number of chains.
}
\description{
Convenience function to extract a log-likelihood matrix or array from the
fitted Stan model in an \code{icfit} object.
}
\seealso{
\code{\link[loo]{extract_log_lik}}
}
