% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\alias{invlogit}
\title{Logit calculations
Helper function to calculate the logit and inverse-logit transforms.}
\usage{
logit(x)

invlogit(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A numeric vector the same length as \code{x}.
}
\description{
Logit calculations
Helper function to calculate the logit and inverse-logit transforms.
}
\keyword{internal}
