% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{loo.icfit}
\alias{loo.icfit}
\alias{loo}
\alias{loo.iclist}
\title{Leave-one-out cross-validation}
\usage{
\method{loo}{icfit}(X, ...)

\method{loo}{iclist}(X, ...)
}
\arguments{
\item{X}{Object of type \code{icfit} or \code{iclist}}

\item{...}{Additional parameters used by \code{rstan::loo}.}
}
\value{
an object of class \code{loo} with the PSIS results.
}
\description{
The \code{loo} method for \code{icfit} objects. computes the approximate
leave-one-out cross-validation using Pareto smoothed importance sampling
(PSIS-LOO CV).
}
\seealso{
\code{\link[loo]{loo}}, \code{\link[rstan]{loo}}
}
