% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{make_ic_svy}
\alias{make_ic_svy}
\title{Internal processing step for ic_survey}
\usage{
make_ic_svy(
  X,
  drop_cols,
  survey,
  sample,
  evidence,
  validity,
  reduce,
  minSample
)
}
\description{
Internal processing step for ic_survey
}
\keyword{internal}
