% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{merge.ic.df}
\alias{merge.ic.df}
\title{Merge two ic data objects
Merge two \code{ic.df} data objects by common columns while preserving the
attributes}
\usage{
\method{merge}{ic.df}(x, y, attr.x = TRUE, ...)
}
\arguments{
\item{x, y}{\code{ic.df} objects to merge}

\item{attr.x}{Logical. Should the attributes (i.e. core field names) be based
on the ic object 'x' or 'y' (when \code{FALSE})?}

\item{...}{arguments to be based on \code{merge}.}
}
\description{
Merge two ic data objects
Merge two \code{ic.df} data objects by common columns while preserving the
attributes
}
