% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{multi_lik_stan}
\alias{multi_lik_stan}
\title{Multi-source immunisation coverage model with Stan}
\usage{
multi_lik_stan(
  X,
  prior_lambda = NULL,
  prior_sigma = NULL,
  upper_sigma = NULL,
  lower_sigma = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{X}{Object of \code{ic.df} for analysis}

\item{prior_lambda}{Scale parameter for half-normal prior on source-specific
intercepts. See details.}

\item{prior_sigma}{Scale parameter for half-Cauchy prior on source-specific
standard deviations. See details.}

\item{upper_sigma}{Numeric value (or vector) setting the upper bounds on the
source-specific scale parameter. See details.}

\item{lower_sigma}{Numeric value (or vector) setting the lower bounds on the
source-specific scale parameter. See details.}

\item{verbose}{Logical. Should messages be displayed? Default is \code{TRUE}.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}.
}
\description{
Multi-source immunisation coverage model with Stan
}
\keyword{internal}
