% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{names}
\alias{names}
\alias{names<-.ic.df}
\alias{names.ic.df}
\title{The names of an ic object}
\usage{
\method{names}{ic.df}(x) <- value

\method{names}{ic.df}(x)
}
\arguments{
\item{x}{An \code{ic.df} object.}

\item{value}{A character vector of up to the same length as \code{x} or NULL.}
}
\value{
For \code{names} a character vector the same length as \code{x}. For
\code{names<-}, the updated ic data frame object is returned.
}
\description{
Functions to get or set the names of an \code{ic.df}.
}
\details{
These access and replacement functions which follow the same
functionality as \code{base::names} for data frames.
}
\seealso{
\code{ic_data}
}
