% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.iclist}
\alias{plot.iclist}
\title{Plot ic fit objects}
\usage{
\method{plot}{iclist}(X, ...)
}
\arguments{
\item{X}{Object of type \code{icfit} or \code{iclist}}

\item{...}{Additional arguments passed to \code{ic_plot}.}
}
\value{
Plot
}
\description{
Provides methods to plot estimated time trends of vaccine coverage for
countries.
}
\seealso{
\code{\link[imcover]{ic_plot}}
}
