% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_interval.R
\name{posterior_interval.icfit}
\alias{posterior_interval.icfit}
\alias{posterior_interval}
\title{Summarise posterior estimates of immunisation coverage}
\usage{
\method{posterior_interval}{icfit}(
  X,
  pars = "mu",
  stat = c("mean", "median", "sd", "quantile"),
  prob = c(0.025, 0.25, 0.5, 0.75, 0.975)
)
}
\arguments{
\item{X}{A fitted model object of type \code{icfit} or \code{iclist}.}

\item{pars}{The parameters to calculate the intervals for.}

\item{stat}{Character vector of summary statistics to apply.}

\item{prob}{Numeric vector of probabilities with values in \link{0,1} to be used
for \code{stat} "quantile".}
}
\value{
A \code{data.frame} with columns for 'country', 'time', and 'vaccine'
labels and summary statistics, and as many rows as there are 'mu'
observations of immunisation coverage.
}
\description{
Computes the Bayesian posterior summary intervals, sometimes
called 'credible intervals' for the estimated immunisation coverage term
('mu').
}
