% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.icfit}
\alias{predict.icfit}
\alias{predict}
\title{Posterior predictions of vaccination coverage}
\usage{
\method{predict}{icfit}(X, country, vaccine, t = 2, return_ic = TRUE)
}
\arguments{
\item{X}{object of type \code{icfit} or \code{iclist}}

\item{country}{character vector of country codes to predict}

\item{vaccine}{character vector of vaccine abbreviations to predict}

\item{t}{Integer of the number of time steps ahead to predict. Default is 2.}

\item{return_ic}{Logical. Should an \code{icfit} object be returned?}
}
\value{
If \code{return_ic} is \code{FALSE}, then a \code{data.frame} with
posterior samples for predictions, labelled by 'country', 'vaccine', and
'time' columns. Return and 'ic' object will modify object \code{X} and
add/update a the 'prediction' element.
}
\description{
Draw from the posterior predictive distribution.
}
