% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{rbind}
\alias{rbind}
\alias{rbind.ic.df}
\title{Combine ic data objects by rows}
\usage{
\method{rbind}{ic.df}(..., fill = TRUE, deparse.level = 1)
}
\arguments{
\item{...}{Objects to bind together.}

\item{fill}{Should missing columns be added to data objects so that they can
be combined? Default is \code{TRUE}.}

\item{deparse.level}{See \code{\link{rbind}} Default is 1.}
}
\value{
An object of type \code{ic.df}.
}
\description{
Take a sequence of \code{ic.df} data-frame arguments and combine by rows.
}
\details{
The attributes of the ic data are used to match columns. Column
names are taken from the first object.
}
