% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.icfit}
\alias{summary.icfit}
\alias{summary.iclist}
\alias{print.summary.icfit}
\alias{print.summary.iclist}
\title{Summary methods for 'ic' fit objects}
\usage{
\method{summary}{icfit}(X, pars = NULL, probs = c(0.1, 0.5, 0.9), ..., digits = 2)

\method{summary}{iclist}(X, ...)

\method{print}{summary.icfit}(X, digits = max(1, attr(X, "print.digits")), ...)

\method{print}{summary.iclist}(X, digits = max(1, attr(X, "print.digits")), ...)
}
\arguments{
\item{X}{An object of class \code{"summary.iclist"}.}

\item{pars}{Optional character vector specifying a subset of parameters to
display.}

\item{probs}{Optional numeric vector of probability passed to
\code{\link[stats]{quantile}}.}

\item{...}{Currently ignored.}

\item{digits}{Number of digits to use for formatting numbers when printing.}
}
\value{
The \code{summary} method returns an object of class
\code{summary.icfit}, which is a matrix of summary statistics and
diagnostics, with attributes storing information for use by the
\code{print} method. The \code{print} method for \code{summary.icfit}
objects is called for its side effect and returns its input.
}
\description{
Summaries of paramter estimates and MCMC convergence diagnostics (Monte Carlo
error, effective sample size, Rhat).
}
