% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_adjust}
\alias{survey_adjust}
\title{Survey recall-bias adjustment}
\usage{
survey_adjust(X, adjVacc = c("DTP", "PCV"))
}
\arguments{
\item{X}{Object of type \code{ic.df}}

\item{adjVacc}{Character vector of vaccine abbreviations (omitting dose
numbers) to adjust.}
}
\value{
An \code{ic.df} object with adjusted coverage estimates for selected
vaccines.
}
\description{
Apply a bias adjustment for selected vaccines in an immunisation coverage
survey dataset.
}
\seealso{
\code{\link[imcover]{ic_survey}}
}
