% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic.icfit}
\alias{waic.icfit}
\alias{waic}
\alias{waic.iclist}
\title{Widely applicable information criterion (WAIC)}
\usage{
\method{waic}{icfit}(X, ...)

\method{waic}{iclist}(X, ...)
}
\arguments{
\item{X}{Object of type \code{icfit} or \code{iclist}.}

\item{...}{Additional arguments. See \code{loo::waic()}.}
}
\value{
A named list (of class \code{c("waic", "loo")}) with components:

\describe{
\item{\code{estimates}}{
A matrix with two columns (\code{"Estimate"}, \code{"SE"}) and three
rows (\code{"elpd_waic"}, \code{"p_waic"}, \code{"waic"}). This contains
point estimates and standard errors of the expected log pointwise predictive
density (\code{elpd_waic}), the effective number of parameters
(\code{p_waic}) and the information criterion \code{waic} (which is just
\code{-2 * elpd_waic}, i.e., converted to deviance scale).
}
\item{\code{pointwise}}{
A matrix with three columns (and number of rows equal to the number of
observations) containing the pointwise contributions of each of the above
measures (\code{elpd_waic}, \code{p_waic}, \code{waic}).
}
}
}
\description{
Implementation of \code{waic()} methods from \code{loo} to compute WAIC from the
pointwise log-likelihood.
}
\details{
For more details see \pkg{loo}. The developers of \code{stan}
recommend LOO-CV using PSIS (as implemented by the \code{loo::loo()}
function) because PSIS provides useful diagnostics as well as effective
sample size and Monte Carlo estimates.
}
\references{
Watanabe, S. (2010). Asymptotic equivalence of Bayes cross validation and
widely application information criterion in singular learning theory.
\emph{Journal of Machine Learning Research} \strong{11}, 3571-3594.
}
\seealso{
\code{\link[loo]{waic}}
}
