% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{as.ic_data}
\alias{as.ic_data}
\alias{is.ic_data}
\title{Convert foreign object to an ic data object}
\usage{
as.ic_data(...)

is.ic_data(object)
}
\arguments{
\item{...}{Elements to be combined into an \code{ic.df} object.}

\item{object}{Any \code{R} object.}
}
\description{
Functions to check if an object is an ic data object, or coerce
it to one if possible.
}
\details{
Core \code{ic.df} data elements are expected. Specifically, data for
'group', 'time', 'vaccine', (optionally 'dose', 'population'), and
'coverage' are required in this order. The names of these elements from
\code{...} are passed to \code{ic_data}.
}
\seealso{
\code{ic_data}
}
