% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_attr}
\alias{get_attr}
\title{Get the attributes of ic data
Retrieve the value(s) of attributes from an \code{ic.df} dataset. While
intended to be used with the core ic attributes, it will work with any named
attribute.}
\usage{
get_attr(X, attrs = ic_core(), unlist = TRUE)
}
\arguments{
\item{X}{An object of class \code{ic.df}.}

\item{attrs}{A vector of character names of attributes to be retrieved.
Default are all core attributes.}

\item{unlist}{Should the function return a named list of attributes and
values or a named vector? Default is \code{TRUE} to return a vector.}
}
\description{
Get the attributes of ic data
Retrieve the value(s) of attributes from an \code{ic.df} dataset. While
intended to be used with the core ic attributes, it will work with any named
attribute.
}
