% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ic_core}
\alias{ic_core}
\title{List core attributes for ic data
Retrieve the names of the core attributes that can be expected in
\code{ic.df} data.}
\usage{
ic_core(survey = FALSE)
}
\arguments{
\item{survey}{Include survey-specific attributes or only \code{ic.df}
attributes? Default is \code{FALSE}.}
}
\description{
List core attributes for ic data
Retrieve the names of the core attributes that can be expected in
\code{ic.df} data.
}
