% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{ic_expand}
\alias{ic_expand}
\title{Expand ic data to include all potential years.}
\usage{
ic_expand(X, min = 1999, max, na.remove = TRUE)
}
\arguments{
\item{X}{Object of class \code{ic.df} to expand.}

\item{min}{Start of the potential time period. Default is 1999.}

\item{max}{End of the potential time period.}

\item{na.remove}{Should country-vaccine combinations which have no coverage
data be removed? Default is \code{TRUE}}
}
\value{
An object of type \code{ic.df} which has additional rows of \code{NA}
values added to expand the dataset to min and max time points.
}
\description{
Expand the time dimension of an ic dataset to include the full set of
potential observations.
}
\details{
If \code{max} is missing, the maximum time value observed in the
dataset will be used.
}
