% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{ic_fit}
\alias{ic_fit}
\alias{ic_fit.ic.df}
\title{Multi-source immunisation coverage model with Stan}
\usage{
ic_fit(X, region = TRUE, verbose = TRUE, ...)

\method{ic_fit}{ic.df}(X, region = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{X}{Object of \code{ic.df} for analysis}

\item{region}{Logical. Should region-specific models be generated?. Default
is \code{TRUE}.}

\item{verbose}{Logical. Should messages be displayed? Default is \code{TRUE}.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Multi-source immunisation coverage model with Stan
}
