% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ic_plot}
\alias{ic_plot}
\alias{ic_plot.icfit}
\alias{ic_plot.iclist}
\title{Plot ic fit objects}
\usage{
ic_plot(
  X,
  probs = c(0.025, 0.975),
  vaccine,
  observed = TRUE,
  prediction = TRUE,
  filter_yovi = FALSE,
  ncol = 4,
  interactive = FALSE
)

\method{ic_plot}{icfit}(
  X,
  probs = c(0.025, 0.975),
  vaccine,
  observed = TRUE,
  prediction = TRUE,
  filter_yovi = FALSE,
  ncol = 4,
  interactive = FALSE
)

\method{ic_plot}{iclist}(
  X,
  probs = c(0.025, 0.975),
  vaccine,
  observed = TRUE,
  prediction = TRUE,
  filter_yovi = FALSE,
  ncol = 4,
  interactive = FALSE
)
}
\arguments{
\item{X}{Object of type \code{icfit} or \code{iclist}}

\item{probs}{Numeric vector of length 2 for probabilities with values in \link{0,1}
to be used to calculate the upper and lower credible intervals.}

\item{vaccine}{Optional. Character vector to subset which vaccines to plot}

\item{observed}{Logical. Should the observed data used to fit \code{X} be
overlaid on the plot? Default is \code{TRUE}.}

\item{prediction}{Logical. If predictions are found in the \code{icfit}
object, should they be overlaid on the plot? Default is \code{TRUE}.}

\item{filter_yovi}{Logical. Should the estimates be filtered by year of
vaccine introduction? Default is \code{FALSE}.}

\item{ncol}{When plotting multiple countries, the number of plots per page.}

\item{interactive}{Logical. Should the output be interactive? Default is
\code{FALSE}.}
}
\value{
A plot of the coverage estimates extracted from the fit, including
credible intervals.
}
\description{
Provides methods to plot estimated time trends of vaccine coverage for
countries.
}
\details{
If observed data are provided, they will be overlaid on the
estimates.
}
