% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{ic_survey}
\alias{ic_survey}
\alias{ic_survey.ic.df}
\alias{ic_survey.data.frame}
\title{Create an ic data object}
\usage{
ic_survey(
  X,
  ...,
  drop_cols = FALSE,
  survey = "surveyNameEnglish",
  sample = "Sample_Size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300
)

\method{ic_survey}{ic.df}(
  X,
  ...,
  drop_cols = FALSE,
  survey = "surveyNameEnglish",
  sample = "Sample_Size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300
)

\method{ic_survey}{ic.df}(
  X,
  ...,
  drop_cols = FALSE,
  survey = "surveyNameEnglish",
  sample = "Sample_Size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300
)

\method{ic_survey}{data.frame}(
  X,
  ...,
  drop_cols = FALSE,
  survey = "surveyNameEnglish",
  sample = "Sample_Size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300
)
}
\arguments{
\item{X}{Any object with immunisation coverage survey data to convert into an
ic object.}

\item{...}{Additional parameters for names of core data elements to pass on
to \code{ic_data}. See details and \code{ic_data}.}

\item{drop_cols}{Should other columns in \code{X} be dropped if they are not
core attributes? Default is \code{FALSE} to retain all data from \code{X}.}

\item{survey, sample, evidence, validity}{Character of the column name within
\code{X} which defines the core values for ic survey data.}

\item{reduce}{Should country-vaccine combinations which have insufficient
coverage data be removed? Default is \code{TRUE}.}

\item{minSample}{Numeric. Minimum reported sample size needed to keep
coverage records. Default is 300. Only used with \code{reduce} is
\code{TRUE}.}
}
\value{
An object of class \code{ic.df} which extends \code{data.frame} with
attributes to location and preserve core data elements for immunisation
coverage.
}
\description{
Create an ic data object from household survey data on
immunisation coverage.
}
\details{
\code{ic_survey} is is an extension to \code{ic_data} and is used to
create a properly formed and processed dataset for immunisation coverage
modelling based on household survey datasets.
}
\examples{
\dontrun{
# assume `df` is a data.frame
# convert to an imcover data frame
ic_survey(df,
         group="iso3",  # specify columns names in df
         time="cohortyear",
         vaccine="vaccine",
         coverage = "coverage",
         survey = "surveynameenglish",
         sample = "sample_size",
         evidence = "evidence",
         validity = "validity",
         biasAdjust = TRUE)
}


}
\seealso{
\code{\link[imcover]{ic_data}}
}
