% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{ic_survey}
\alias{ic_survey}
\alias{ic_survey.ic.svy}
\alias{ic_survey.ic.df}
\alias{ic_survey.data.frame}
\title{Create an ic data object}
\usage{
ic_survey(
  X,
  ...,
  dropCols = FALSE,
  survey = "survey_name",
  sample = "sample_size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300,
  expand = TRUE,
  min = 1999,
  max,
  na.remove = TRUE,
  biasAdjust = TRUE,
  adjVacc = c("DTP", "PCV")
)

\method{ic_survey}{ic.svy}(
  X,
  ...,
  dropCols = FALSE,
  survey = "survey_name",
  sample = "sample_size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300,
  expand = TRUE,
  min = 1999,
  max,
  na.remove = TRUE,
  biasAdjust = TRUE,
  adjVacc = c("DTP", "PCV")
)

\method{ic_survey}{ic.df}(
  X,
  ...,
  dropCols = FALSE,
  survey = "survey_name",
  sample = "sample_size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300,
  expand = TRUE,
  min = 1999,
  max,
  na.remove = TRUE,
  biasAdjust = TRUE,
  adjVacc = c("DTP", "PCV")
)

\method{ic_survey}{data.frame}(
  X,
  ...,
  dropCols = FALSE,
  survey = "survey_name",
  sample = "sample_size",
  evidence = "evidence",
  validity = "validity",
  reduce = TRUE,
  minSample = 300,
  expand = TRUE,
  min = 1999,
  max,
  na.remove = TRUE,
  biasAdjust = TRUE,
  adjVacc = c("DTP", "PCV")
)
}
\arguments{
\item{X}{Any object with immunisation coverage survey data to convert into an
ic object.}

\item{...}{Additional parameters for names of core data elements to pass on
to \code{ic_data}. See details and \code{ic_data}.}

\item{survey, sample, evidence, validity}{Character of the name within \code{X}
which defines the core value for ic survey data.}

\item{reduce}{Should group-vaccine combinations which have no coverage data
be removed? Default is \code{TRUE}.}

\item{minSample}{Numeric. Minimum reported sample size. Default is 300.}

\item{expand}{Should additional empty rows be added to expand the ic data to
all possible group x vaccine x time observation? Default is \code{FALSE}.}

\item{min}{Start of the potential time period. Default is 1999.}

\item{max}{End of the potential time period.}
}
\value{
An object of class \code{ic.svy} which extends \code{ic.df} and
\code{data.frame}-like objects with attributes to location and preserve
core data elements for immunisation coverage.
}
\description{
Create an ic data object from survey data on immunisation
coverage.
}
\details{
\code{ic_survey} is is an extension to \code{ic_data} and is used to
create a properly formed and processed dataset for immunisation coverage
modelling based on survey datasets.

The parameters \code{min} and \code{max} are only needed when \code{expand =
TRUE}. If \code{max} is missing, the maximum time value observed in the
dataset will be used.
}
\seealso{
\code{\link[imcover]{ic_data}}
}
