% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imcover.R
\docType{package}
\name{imcover-package}
\alias{imcover-package}
\title{imcover: An R package for immunisation coverage modelling}
\description{
Provides access to WUENIC immunisation data and a Bayesian statistical
modelling framework to produces country- and region-specific temporal trends
in immunisation coverage.
}
\section{Credits}{

This work was undertaken by members of the WorldPop Research Group at the
University of Southampton(\url{https://www.worldpop.org/}).
}

