% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{list_vaccines}
\alias{list_vaccines}
\title{List vaccines}
\usage{
list_vaccines(X)
}
\arguments{
\item{X}{Object of class \code{ic.df}.}
}
\description{
Retrieve the unique list of vaccines observed in an ic dataset.
}
