% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{multi_lik_stan}
\alias{multi_lik_stan}
\title{Multi-source immunisation coverage model with Stan}
\usage{
multi_lik_stan(x, y, ...)
}
\arguments{
\item{x}{Numeric vector of input values.}

\item{y}{Numberic vector of output values.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Multi-source immunisation coverage model with Stan
}
