% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan.R
\name{multi_lik_stan}
\alias{multi_lik_stan}
\title{Multi-source immunisation coverage model with Stan}
\usage{
multi_lik_stan(X, verbose = TRUE, ...)
}
\arguments{
\item{X}{Object of \code{ic.df} for analysis}

\item{verbose}{Logical. Should messages be displayed? Default is \code{TRUE}.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Multi-source immunisation coverage model with Stan
}
