% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.icfit}
\alias{predict.icfit}
\alias{predict}
\title{Posterior predictions of vaccination coverage}
\usage{
\method{predict}{icfit}(
  X,
  country = NULL,
  vaccine = NULL,
  t = 2,
  ratio_adj = TRUE,
  return_ic = TRUE
)
}
\arguments{
\item{X}{object of type \code{icfit} or \code{iclist}}

\item{country}{character vector of country codes to predict}

\item{vaccine}{character vector of vaccine abbreviations to predict}

\item{t}{Integer of the number of time steps ahead to predict. Default is 2.}

\item{ratio_adj}{Logical. Should multi-dose vaccines modelled by a ratio be
back-transformed to coverage? Default is \code{TRUE}.}

\item{return_ic}{Logical. Should an \code{icfit} object be returned?}
}
\value{
If \code{return_ic} is \code{FALSE}, then a \code{data.frame} (or a
list of data frames) with posterior samples for predictions, labelled by
'country', 'vaccine', and 'time' columns. Returning an 'ic' object will
modify object \code{X} and add/update with a the 'prediction' element and
return the same type as \code{X}.
}
\description{
Draw from the posterior predictive distribution.
}
